+++ lib/edit_address.dart
@@ -1,3 +1,4 @@
+/// {@category Edit Address}
 library edit_address;
 
 export 'src/lib_edit_address/models/address_model.dart';
@@ -7,5 +8,5 @@ export 'src/lib_edit_address/models/user_profile_card_types.dart';
 export 'src/lib_edit_address/resources/edit_address_localized_strings.dart';
 export 'src/lib_edit_address/services/edit_address_service.dart';
 export 'src/lib_edit_address/theme/edit_address_theme.dart';
-export 'src/lib_edit_address/ui_components/address_for_correspondence_bottom_sheet.dart';
+export 'src/lib_edit_address/ui_components/edit_address_bottom_sheet.dart';
 export 'src/lib_edit_address/ui_components/edit_address_widget.dart';