+++ lib/src/base/theme/widget_toolkit_theme.tailor.dart
@@ -56,6 +56,7 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
     required this.filledButtonBackgroundColorDisabled,
     required this.filledButtonTextColorDisabled,
     required this.textButtonTextColorDisabled,
+    required this.textButtonBackgroundColor,
     required this.filledButtonTextColorEnabled,
     required this.smallButtonBackgroundColor,
     required this.buttonTextColor,
@@ -94,6 +95,7 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
     required this.outlineButtonPressedColor,
     required this.outlineButtonTextColor,
     required this.outlineButtonTextColorDisabled,
+    required this.bottomSheetCloseButtonPadding,
     required this.messagePanelErrorEdgeInsets,
     required this.smallEdgeInsets,
     required this.mediumEdgeInsets,
@@ -166,6 +168,7 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
   final Color filledButtonBackgroundColorDisabled;
   final Color filledButtonTextColorDisabled;
   final Color textButtonTextColorDisabled;
+  final Color textButtonBackgroundColor;
   final Color filledButtonTextColorEnabled;
   final Color smallButtonBackgroundColor;
   final Color buttonTextColor;
@@ -204,6 +207,7 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
   final Color outlineButtonPressedColor;
   final Color outlineButtonTextColor;
   final Color outlineButtonTextColorDisabled;
+  final EdgeInsets bottomSheetCloseButtonPadding;
   final EdgeInsets messagePanelErrorEdgeInsets;
   final EdgeInsets smallEdgeInsets;
   final EdgeInsets mediumEdgeInsets;
@@ -293,6 +297,8 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
         _$WidgetToolkitTheme.filledButtonTextColorDisabled[0],
     textButtonTextColorDisabled:
         _$WidgetToolkitTheme.textButtonTextColorDisabled[0],
+    textButtonBackgroundColor:
+        _$WidgetToolkitTheme.textButtonBackgroundColor[0],
     filledButtonTextColorEnabled:
         _$WidgetToolkitTheme.filledButtonTextColorEnabled[0],
     smallButtonBackgroundColor:
@@ -344,6 +350,8 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
     outlineButtonTextColor: _$WidgetToolkitTheme.outlineButtonTextColor[0],
     outlineButtonTextColorDisabled:
         _$WidgetToolkitTheme.outlineButtonTextColorDisabled[0],
+    bottomSheetCloseButtonPadding:
+        _$WidgetToolkitTheme.bottomSheetCloseButtonPadding[0],
     messagePanelErrorEdgeInsets:
         _$WidgetToolkitTheme.messagePanelErrorEdgeInsets[0],
     smallEdgeInsets: _$WidgetToolkitTheme.smallEdgeInsets[0],
@@ -436,6 +444,8 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
         _$WidgetToolkitTheme.filledButtonTextColorDisabled[1],
     textButtonTextColorDisabled:
         _$WidgetToolkitTheme.textButtonTextColorDisabled[1],
+    textButtonBackgroundColor:
+        _$WidgetToolkitTheme.textButtonBackgroundColor[1],
     filledButtonTextColorEnabled:
         _$WidgetToolkitTheme.filledButtonTextColorEnabled[1],
     smallButtonBackgroundColor:
@@ -487,6 +497,8 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
     outlineButtonTextColor: _$WidgetToolkitTheme.outlineButtonTextColor[1],
     outlineButtonTextColorDisabled:
         _$WidgetToolkitTheme.outlineButtonTextColorDisabled[1],
+    bottomSheetCloseButtonPadding:
+        _$WidgetToolkitTheme.bottomSheetCloseButtonPadding[1],
     messagePanelErrorEdgeInsets:
         _$WidgetToolkitTheme.messagePanelErrorEdgeInsets[1],
     smallEdgeInsets: _$WidgetToolkitTheme.smallEdgeInsets[1],
@@ -568,6 +580,7 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
     Color? filledButtonBackgroundColorDisabled,
     Color? filledButtonTextColorDisabled,
     Color? textButtonTextColorDisabled,
+    Color? textButtonBackgroundColor,
     Color? filledButtonTextColorEnabled,
     Color? smallButtonBackgroundColor,
     Color? buttonTextColor,
@@ -606,6 +619,7 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
     Color? outlineButtonPressedColor,
     Color? outlineButtonTextColor,
     Color? outlineButtonTextColorDisabled,
+    EdgeInsets? bottomSheetCloseButtonPadding,
     EdgeInsets? messagePanelErrorEdgeInsets,
     EdgeInsets? smallEdgeInsets,
     EdgeInsets? mediumEdgeInsets,
@@ -708,6 +722,8 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
           filledButtonTextColorDisabled ?? this.filledButtonTextColorDisabled,
       textButtonTextColorDisabled:
           textButtonTextColorDisabled ?? this.textButtonTextColorDisabled,
+      textButtonBackgroundColor:
+          textButtonBackgroundColor ?? this.textButtonBackgroundColor,
       filledButtonTextColorEnabled:
           filledButtonTextColorEnabled ?? this.filledButtonTextColorEnabled,
       smallButtonBackgroundColor:
@@ -768,6 +784,8 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
           outlineButtonTextColor ?? this.outlineButtonTextColor,
       outlineButtonTextColorDisabled:
           outlineButtonTextColorDisabled ?? this.outlineButtonTextColorDisabled,
+      bottomSheetCloseButtonPadding:
+          bottomSheetCloseButtonPadding ?? this.bottomSheetCloseButtonPadding,
       messagePanelErrorEdgeInsets:
           messagePanelErrorEdgeInsets ?? this.messagePanelErrorEdgeInsets,
       smallEdgeInsets: smallEdgeInsets ?? this.smallEdgeInsets,
@@ -896,6 +914,8 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
           other.filledButtonTextColorDisabled, t)!,
       textButtonTextColorDisabled: Color.lerp(
           textButtonTextColorDisabled, other.textButtonTextColorDisabled, t)!,
+      textButtonBackgroundColor: Color.lerp(
+          textButtonBackgroundColor, other.textButtonBackgroundColor, t)!,
       filledButtonTextColorEnabled: Color.lerp(
           filledButtonTextColorEnabled, other.filledButtonTextColorEnabled, t)!,
       smallButtonBackgroundColor: Color.lerp(
@@ -968,6 +988,9 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
           Color.lerp(outlineButtonTextColor, other.outlineButtonTextColor, t)!,
       outlineButtonTextColorDisabled: Color.lerp(outlineButtonTextColorDisabled,
           other.outlineButtonTextColorDisabled, t)!,
+      bottomSheetCloseButtonPadding: t < 0.5
+          ? bottomSheetCloseButtonPadding
+          : other.bottomSheetCloseButtonPadding,
       messagePanelErrorEdgeInsets: t < 0.5
           ? messagePanelErrorEdgeInsets
           : other.messagePanelErrorEdgeInsets,
@@ -1073,6 +1096,7 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
             const DeepCollectionEquality().equals(filledButtonBackgroundColorDisabled, other.filledButtonBackgroundColorDisabled) &&
             const DeepCollectionEquality().equals(filledButtonTextColorDisabled, other.filledButtonTextColorDisabled) &&
             const DeepCollectionEquality().equals(textButtonTextColorDisabled, other.textButtonTextColorDisabled) &&
+            const DeepCollectionEquality().equals(textButtonBackgroundColor, other.textButtonBackgroundColor) &&
             const DeepCollectionEquality().equals(filledButtonTextColorEnabled, other.filledButtonTextColorEnabled) &&
             const DeepCollectionEquality().equals(smallButtonBackgroundColor, other.smallButtonBackgroundColor) &&
             const DeepCollectionEquality().equals(buttonTextColor, other.buttonTextColor) &&
@@ -1111,6 +1135,7 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
             const DeepCollectionEquality().equals(outlineButtonPressedColor, other.outlineButtonPressedColor) &&
             const DeepCollectionEquality().equals(outlineButtonTextColor, other.outlineButtonTextColor) &&
             const DeepCollectionEquality().equals(outlineButtonTextColorDisabled, other.outlineButtonTextColorDisabled) &&
+            const DeepCollectionEquality().equals(bottomSheetCloseButtonPadding, other.bottomSheetCloseButtonPadding) &&
             const DeepCollectionEquality().equals(messagePanelErrorEdgeInsets, other.messagePanelErrorEdgeInsets) &&
             const DeepCollectionEquality().equals(smallEdgeInsets, other.smallEdgeInsets) &&
             const DeepCollectionEquality().equals(mediumEdgeInsets, other.mediumEdgeInsets) &&
@@ -1187,6 +1212,7 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
       const DeepCollectionEquality().hash(filledButtonBackgroundColorDisabled),
       const DeepCollectionEquality().hash(filledButtonTextColorDisabled),
       const DeepCollectionEquality().hash(textButtonTextColorDisabled),
+      const DeepCollectionEquality().hash(textButtonBackgroundColor),
       const DeepCollectionEquality().hash(filledButtonTextColorEnabled),
       const DeepCollectionEquality().hash(smallButtonBackgroundColor),
       const DeepCollectionEquality().hash(buttonTextColor),
@@ -1225,6 +1251,7 @@ class WidgetToolkitTheme extends ThemeExtension<WidgetToolkitTheme> {
       const DeepCollectionEquality().hash(outlineButtonPressedColor),
       const DeepCollectionEquality().hash(outlineButtonTextColor),
       const DeepCollectionEquality().hash(outlineButtonTextColorDisabled),
+      const DeepCollectionEquality().hash(bottomSheetCloseButtonPadding),
       const DeepCollectionEquality().hash(messagePanelErrorEdgeInsets),
       const DeepCollectionEquality().hash(smallEdgeInsets),
       const DeepCollectionEquality().hash(mediumEdgeInsets),