+++ lib/src/lib_edit_address/views/edit_address_page.dart
@@ -4,14 +4,19 @@ import 'package:flutter_rx_bloc/rx_form.dart';
 import 'package:provider/provider.dart';
 
 import '../../../edit_address.dart';
+import '../../../models.dart';
 import '../../../search_picker.dart';
-import '../../../widget_toolkit.dart';
-import '../../lib_edit_address/ui_components/edit_address_form.dart';
+import '../../../text_field_dialog.dart';
+import '../../../ui_components.dart';
 import '../blocs/edit_address_bloc.dart';
 import '../di/edit_address_dependencies.dart';
+import '../ui_components/edit_address_form.dart';
+
+typedef OnAddressSaved = Function(AddressModel addressModel);
 
 class EditAddressPage<T extends PickerItemModel> extends StatelessWidget {
   const EditAddressPage({
+    required this.onAddressSaved,
     required this.addressModel,
     required this.cityErrorMapper,
     required this.addressErrorMapper,
@@ -30,9 +35,13 @@ class EditAddressPage<T extends PickerItemModel> extends StatelessWidget {
     this.editAddressFieldType = EditFieldType.editfield,
     this.editContactAddressErrorBuilder,
     this.searchCountryCustomBuilders,
+    this.textFieldsModalConfiguration = const TextFieldModalConfiguration(),
+    this.countryPickerModalConfiguration =
+        const SearchPickerModalConfiguration(),
     Key? key,
   }) : super(key: key);
 
+  final OnAddressSaved onAddressSaved;
   final AddressModel addressModel;
   final String buttonText;
   final String headerText;
@@ -53,9 +62,12 @@ class EditAddressPage<T extends PickerItemModel> extends StatelessWidget {
   final EditAddressService editAddressService;
   final Widget Function(ErrorModel?)? editContactAddressErrorBuilder;
   final SearchCountryCustomBuilders<T>? searchCountryCustomBuilders;
+  final TextFieldModalConfiguration textFieldsModalConfiguration;
+  final SearchPickerModalConfiguration countryPickerModalConfiguration;
 
   static Widget withDependencies<T extends PickerItemModel>(
     BuildContext context, {
+    required OnAddressSaved onAddressSaved,
     required String buttonText,
     required String headerText,
     required AddressModel addressModel,
@@ -78,6 +90,10 @@ class EditAddressPage<T extends PickerItemModel> extends StatelessWidget {
     final EditFieldType editAddressFieldType = EditFieldType.editfield,
     final Widget Function(ErrorModel?)? editContactAddressErrorBuilder,
     final SearchCountryCustomBuilders<T>? searchCountryCustomBuilders,
+    final TextFieldModalConfiguration textFieldsModalConfiguration =
+        const TextFieldModalConfiguration(),
+    final SearchPickerModalConfiguration countryPickerModalConfiguration =
+        const SearchPickerModalConfiguration(),
   }) =>
       MultiProvider(
         providers: EditAddressDependencies.from(
@@ -86,24 +102,26 @@ class EditAddressPage<T extends PickerItemModel> extends StatelessWidget {
           editAddressService,
         ).providers,
         child: EditAddressPage<T>(
-          buttonText: buttonText,
-          headerText: headerText,
-          addressModel: addressModel,
-          saveAddress: saveAddress,
-          cityErrorMapper: cityErrorMapper,
-          addressErrorMapper: addressErrorMapper,
-          validator: validator,
-          countryCustomIcon: countryCustomIcon,
-          editCountryFieldType: editCountryFieldType,
-          cityCustomIcon: cityCustomIcon,
-          editCityFieldType: editCityFieldType,
-          addressCustomIcon: addressCustomIcon,
-          editAddressFieldType: editAddressFieldType,
-          searchCountryService: searchCountryService,
-          editAddressLocalizedStrings: editAddressLocalizedStrings,
-          editAddressService: editAddressService,
-          searchCountryCustomBuilders: searchCountryCustomBuilders,
-        ),
+            onAddressSaved: onAddressSaved,
+            buttonText: buttonText,
+            headerText: headerText,
+            addressModel: addressModel,
+            saveAddress: saveAddress,
+            cityErrorMapper: cityErrorMapper,
+            addressErrorMapper: addressErrorMapper,
+            validator: validator,
+            countryCustomIcon: countryCustomIcon,
+            editCountryFieldType: editCountryFieldType,
+            cityCustomIcon: cityCustomIcon,
+            editCityFieldType: editCityFieldType,
+            addressCustomIcon: addressCustomIcon,
+            editAddressFieldType: editAddressFieldType,
+            searchCountryService: searchCountryService,
+            editAddressLocalizedStrings: editAddressLocalizedStrings,
+            editAddressService: editAddressService,
+            searchCountryCustomBuilders: searchCountryCustomBuilders,
+            textFieldsModalConfiguration: textFieldsModalConfiguration,
+            countryPickerModalConfiguration: countryPickerModalConfiguration),
       );
 
   @override
@@ -114,88 +132,90 @@ class EditAddressPage<T extends PickerItemModel> extends StatelessWidget {
           child: Container(
             padding: context.editAddressTheme.editAddressPageOuterMostPadding,
             color: context.editAddressTheme.editAddressPageBackgroundColor,
-            child: RxBlocBuilder<EditAddressBlocType, void>(
+            child: RxBlocListener<EditAddressBlocType, AddressModel>(
               state: (bloc) => bloc.states.onAddressSaved,
-              builder: (context, snapshot, bloc) => Column(
-                crossAxisAlignment: CrossAxisAlignment.stretch,
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  Padding(
-                    padding: context
-                        .editAddressTheme.editAddressPageOnAddressSavedPadding,
-                    child: Text(
-                      headerText,
-                      textAlign: TextAlign.start,
-                      style: context.editAddressTheme.titleBold,
-                    ),
-                  ),
-                  if (snapshot.hasData)
-                    MessagePanelWidget(
-                      message:
-                          editAddressLocalizedStrings?.addressChangedMessage ??
-                              context.getEditAddressLocalizedStrings
-                                  .addressChangedMessage,
-                      messageState: MessagePanelState.positiveCheck,
-                    ),
-                  MessagePanelError<EditAddressBlocType>(
-                    errorState: (bloc) => bloc.states.errors,
-                    errorBuilder: editContactAddressErrorBuilder,
-                    padding: context
-                        .editAddressTheme.editAddressPageErrorPanelPadding,
-                  ),
-                  if (!snapshot.hasData)
-                    EditAddressForm.withDependencies<T>(
-                      context,
-                      addressModel,
-                      (addressModel) => context
-                          .read<EditAddressBlocType>()
-                          .events
-                          .setAddress(addressModel),
-                      cityErrorMapper: cityErrorMapper,
-                      addressErrorMapper: addressErrorMapper,
-                      validator: validator,
-                      countryCustomIcon: countryCustomIcon,
-                      editCountryFieldType: editCountryFieldType,
-                      cityCustomIcon: cityCustomIcon,
-                      editCityFieldType: editCityFieldType,
-                      addressCustomIcon: addressCustomIcon,
-                      editAddressFieldType: editAddressFieldType,
-                      searchCountryService: searchCountryService,
-                      editAddressLocalizedStrings: editAddressLocalizedStrings,
-                      editAddressService: editAddressService,
-                      searchCountryCustomBuilders: searchCountryCustomBuilders,
-                    ),
-                  if (!snapshot.hasData)
-                    SizedBox(height: context.editAddressTheme.spacingXL),
-                  if (!snapshot.hasData)
-                    RxBlocBuilder<EditAddressBlocType, bool>(
-                      state: (bloc) => bloc.states.isLoading,
-                      builder: (context, isLoadingSnapshot, bloc) =>
-                          RxBlocBuilder<EditAddressBlocType, AddressModel>(
-                        state: (bloc) => bloc.states.onAddressSet,
-                        builder: (context, address, bloc) => GradientFillButton(
-                          elevation: 0,
-                          text: buttonText,
-                          state: isLoadingSnapshot.data ?? false
-                              ? ButtonStateModel.loading
-                              : ButtonStateModel.enabled,
-                          areIconsClose: true,
-                          onPressed: () async {
-                            if (saveAddress == null) {
-                              context
-                                  .read<EditAddressBlocType>()
-                                  .events
-                                  .saveAddress();
-                            } else {
-                              if (address.hasData) {
-                                saveAddress?.call(address.data!);
-                              }
-                            }
-                          },
-                        ),
+              listener: (context, state) => onAddressSaved(state),
+              child: RxBlocBuilder<EditAddressBlocType, AddressModel>(
+                state: (bloc) => bloc.states.onAddressSaved,
+                builder: (context, snapshot, bloc) => Column(
+                  crossAxisAlignment: CrossAxisAlignment.stretch,
+                  mainAxisSize: MainAxisSize.min,
+                  children: [
+                    Padding(
+                      padding: context.editAddressTheme
+                          .editAddressPageOnAddressSavedPadding,
+                      child: Text(
+                        headerText,
+                        textAlign: TextAlign.start,
+                        style: context.editAddressTheme.titleBold,
                       ),
                     ),
-                ],
+                    MessagePanelError<EditAddressBlocType>(
+                      errorState: (bloc) => bloc.states.errors,
+                      errorBuilder: editContactAddressErrorBuilder,
+                      padding: context
+                          .editAddressTheme.editAddressPageErrorPanelPadding,
+                    ),
+                    if (!snapshot.hasData)
+                      EditAddressForm.withDependencies<T>(context, addressModel,
+                          (addressModel) {
+                        return context
+                            .read<EditAddressBlocType>()
+                            .events
+                            .setAddress(addressModel);
+                      },
+                          cityErrorMapper: cityErrorMapper,
+                          addressErrorMapper: addressErrorMapper,
+                          validator: validator,
+                          countryCustomIcon: countryCustomIcon,
+                          editCountryFieldType: editCountryFieldType,
+                          cityCustomIcon: cityCustomIcon,
+                          editCityFieldType: editCityFieldType,
+                          addressCustomIcon: addressCustomIcon,
+                          editAddressFieldType: editAddressFieldType,
+                          searchCountryService: searchCountryService,
+                          editAddressLocalizedStrings:
+                              editAddressLocalizedStrings,
+                          editAddressService: editAddressService,
+                          searchCountryCustomBuilders:
+                              searchCountryCustomBuilders,
+                          textFieldsModalConfiguration:
+                              textFieldsModalConfiguration,
+                          countryPickerModalConfiguration:
+                              countryPickerModalConfiguration),
+                    if (!snapshot.hasData)
+                      SizedBox(height: context.editAddressTheme.spacingXL),
+                    if (!snapshot.hasData)
+                      RxBlocBuilder<EditAddressBlocType, bool>(
+                        state: (bloc) => bloc.states.isLoading,
+                        builder: (context, isLoadingSnapshot, bloc) =>
+                            RxBlocBuilder<EditAddressBlocType, AddressModel>(
+                          state: (bloc) => bloc.states.onAddressSet,
+                          builder: (context, address, bloc) =>
+                              GradientFillButton(
+                            elevation: 0,
+                            text: buttonText,
+                            state: isLoadingSnapshot.data ?? false
+                                ? ButtonStateModel.loading
+                                : ButtonStateModel.enabled,
+                            areIconsClose: true,
+                            onPressed: () async {
+                              if (saveAddress == null) {
+                                context
+                                    .read<EditAddressBlocType>()
+                                    .events
+                                    .saveAddress();
+                              } else {
+                                if (address.hasData) {
+                                  saveAddress?.call(address.data!);
+                                }
+                              }
+                            },
+                          ),
+                        ),
+                      ),
+                  ],
+                ),
               ),
             ),
           ),