+++ lib/src/lib_search_picker/model/item_model.dart
@@ -1,4 +1,0 @@
-//TODO: Include this in item picker improvements
-abstract class ItemModel {
-  String? get itemDisplayName;
-}