+++ lib/src/lib_search_picker/services/search_picker_service.dart
@@ -1,7 +1,17 @@
 abstract class SearchPickerService<T> {
   Future<List<T>> getItems();
 
-  Future<List<T>> filteredListByName(List<T> list, String? searchParam);
+  /// By default this function checks if the item.toString() contains the searchParam.
+  /// Not case-sensitive
+  Future<List<T>> filteredListByName(List<T> list, String? searchParam) async {
+    if (searchParam == null) {
+      return list;
+    }
+    return list
+        .where((item) =>
+            (item.toString()).toLowerCase().contains(searchParam.toLowerCase()))
+        .toList();
+  }
 
   List<T> getPlaceholderList();
 }