+++ lib/ui_components.dart
@@ -1,3 +1,4 @@
+/// {@category UI Components}
 library ui_components;
 
 export 'src/lib_ui_components/buttons/button_color_style.dart';
@@ -10,6 +11,6 @@ export 'src/lib_ui_components/dynamic_icon.dart';
 export 'src/lib_ui_components/error_card_widget.dart';
 export 'src/lib_ui_components/message_panel_error.dart';
 export 'src/lib_ui_components/message_panel_widget.dart';
-export 'src/lib_ui_components/show_error_modal.dart';
-export 'src/lib_ui_components/show_modal.dart';
+export 'src/lib_ui_components/show_blurred_bottom_sheet.dart';
+export 'src/lib_ui_components/show_error_blurred_bottom_sheet.dart';
 export 'src/lib_ui_components/sized_loading_indicator.dart';